fix: isolate investigation constants for client imports
